+++ src/Handlers/acThresholdNow.cpp
@@ -37,7 +37,7 @@ int acThresholdNow::Init () {
 			return -1;
 		}
 	        solver->lattice->SetSetting(it, level);
-	        if (slice != NULL) for (int j=0;j<par;j++) slice[j]=start[j]>level ? 1.0 : 0.0;
+	        if (slice != NULL) for (size_t j=0;j<par;j++) slice[j]=start[j]>level ? 1.0 : 0.0;
 //       		solver->setPar(slice);
 		SetParameters(slice);
 		return 0;
