+++ src/Handlers/conFieldParameter.h
@@ -14,7 +14,7 @@ class  conFieldParameter  : public Design  {
 	int *Par_sizes; ///< Parameter space dimensions on all the processors
 	int *Par_disp; ///< Offsets in the Parameter vector for all the processors
 	int mpi_size, mpi_rank;
-	int CalculateNumberOfParameters ();
+	size_t CalculateNumberOfParameters ();
 	bool FlagInDesignSpace(flag_t);
 	bool InDesignSpace(size_t);
 	int LocalParameters (int type, double * tab);
@@ -22,7 +22,7 @@ class  conFieldParameter  : public Design  {
 public:
 	static std::string xmlname;
 	int Init ();
-	int NumberOfParameters ();
+	size_t NumberOfParameters ();
 	int Parameters (int type, double * tab);
 };
