+++ src/mpitools.hpp
@@ -12,7 +12,7 @@ inline std::string MPI_Bcast(const std::string& str, int root, MPI_Comm comm) {
         ::MPI_Bcast(&size, 1, MPI_UNSIGNED_LONG, root, comm);
         char * buf = new char[size+1];
         strcpy(buf, str.c_str());
-        ::MPI_Bcast(buf, size+1, MPI_CHAR, root, comm);
+        ::MPI_Bcast(buf, (int) size+1, MPI_CHAR, root, comm);
         std::string ret(buf,size);
         delete[] buf;
         return ret;
